Week 13/Week13 day1 - Langgraph 1/requirements.txt
langgraph
langchain-core
langchain-openai
python-dotenv
